# lib/api-zod/src/generated/types/agentChatRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Duplicate Payment Detection System API
 * OpenAPI spec version: 0.1.0
 */
import type { AgentChatRequestAgentType } from "./agentChatRequestAgentType";

export interface AgentChatRequest {
  message: string;
  agentType?: AgentChatRequestAgentType;
  conversationId?: string;
}