# lib/api-zod/src/generated/types/duplicatePayment.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Duplicate Payment Detection System API
 * OpenAPI spec version: 0.1.0
 */

export interface DuplicatePayment {
  id: string;
  payment1Id: string;
  payment2Id: string;
  /** 0.0 to 1.0 probability this is a true duplicate */
  probability: number;
  /** Type of duplicate (exact_match, fuzzy_match, multi_source, etc.) */
  duplicateType: string;
  /** SWIFT_MT, SWIFT_MX, ACH, INTERNAL */
  paymentSystem: string;
  amount: number;
  currency: string;
  senderBIC?: string;
  receiverBIC?: string;
  originatorCountry?: string;
  beneficiaryCountry?: string;
  paymentDate1?: string;
  paymentDate2?: string;
  /** pending, confirmed_duplicate, dismissed, under_review */
  status: string;
  matchedFields?: string[];
  detectedAt: string;
  notes?: string;
}