# lib/api-zod/src/generated/types/exportRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Duplicate Payment Detection System API
 * OpenAPI spec version: 0.1.0
 */
import type { ExportRequestFilters } from "./exportRequestFilters";
import type { ExportRequestFormat } from "./exportRequestFormat";

export interface ExportRequest {
  format: ExportRequestFormat;
  filters?: ExportRequestFilters;
}