# lib/api-zod/src/generated/types/updateDuplicateStatusBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Duplicate Payment Detection System API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateDuplicateStatusBodyStatus } from "./updateDuplicateStatusBodyStatus";

export interface UpdateDuplicateStatusBody {
  status: UpdateDuplicateStatusBodyStatus;
  notes?: string;
}